sales invoice return

+++ app/Http/Controllers/Admin/SalesInvoiceReturnController.php
@@ -5,14 +5,17 @@ namespace App\Http\Controllers\Admin;
 use App\Models\Part;
 use App\Models\PointRule;
 use App\Models\ReturnedSaleReceipt;
+use App\Models\SupplyTerm;
 use App\Models\User;
 use App\Models\Branch;
 use App\Models\Customer;
 use App\Models\TaxesFees;
 use App\Models\SalesInvoice;
+use App\Services\HandleQuantityService;
 use App\Services\MailServices;
 use App\Services\NotificationServices;
 use App\Services\PointsServices;
+use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Http\Request;
 use App\Models\PurchaseInvoice;
 use App\Models\SalesInvoiceItems;
@@ -26,15 +29,17 @@ use Illuminate\Support\Facades\Validator;
 use \App\Services\SalesInvoiceReturnServices;
 use App\Http\Controllers\ExportPrinterFactory;
 use App\Http\Controllers\DataExportCore\Invoices\SalesReturn;
-use App\Http\Requests\Admin\SalesInvoicesReturn\CreateSalesInvoiceReturn;
 use App\Http\Requests\Admin\SalesInvoicesReturn\CreateSalesInvoiceReturnRequest;
 use App\Http\Requests\Admin\SalesInvoicesReturn\UpdateSalesInvoiceReturnRequest;
+use Yajra\DataTables\DataTables;
 
 class SalesInvoiceReturnController extends Controller
 {
-    use  SalesInvoiceReturnServices, NotificationServices, MailServices, PointsServices;
+    use   NotificationServices, MailServices;
 
     public $lang;
+    public $salesInvoiceReturn;
+    public $handleQuantityServices;
 
     public function __construct()
     {
@@ -44,21 +49,41 @@ class SalesInvoiceReturnController extends Controller
 //        $this->middleware('permission:delete_sales_invoices_return',['only'=>['destroy','deleteSelected']]);
 
         $this->lang = App::getLocale();
-
+        $this->salesInvoiceReturn = new SalesInvoiceReturnServices();
+        $this->handleQuantityServices = new HandleQuantityService();
     }
 
     public function index(Request $request)
     {
+        if (!auth()->user()->can('view_sales_invoices_return')) {
+            return redirect()->back()->with(['authorization' => 'error']);
+        }
 
-//        if (!auth()->user()->can('view_sales_invoices_return')) {
-//            return redirect()->back()->with(['authorization' => 'error']);
-//        }
+        $data = SalesInvoiceReturn::query()->latest();
 
-//        $invoices = SalesInvoiceReturn::query();
+        if ($request->filled('filter')) {
+            $data = $this->filter($request, $data);
+        }
 
-        return view('admin.sales_invoice_return.index');
+        $paymentTerms = SupplyTerm::where('purchase_invoice', 1)->where('status', 1)->where('type', 'payment')
+            ->select('id', 'term_' . $this->lang)->get();
 
-//        return view('admin.sales_invoice_return.index',compact('invoices'));
+        $supplyTerms = SupplyTerm::where('purchase_invoice', 1)->where('status', 1)->where('type', 'supply')
+            ->select('id', 'term_' . $this->lang)->get();
+
+        if ($request->isDataTable) {
+
+            return $this->dataTableColumns($data);
+
+        } else {
+
+            return view('admin.sales_invoice_return.index', [
+                'data' => $data,
+                'paymentTerms' => $paymentTerms,
+                'supplyTerms' => $supplyTerms,
+                'js_columns' => SalesInvoiceReturn::getJsDataTablesColumns(),
+            ]);
+        }
     }
 
     public function create(Request $request)
@@ -87,119 +112,69 @@ class SalesInvoiceReturnController extends Controller
         return view('admin.sales_invoice_return.create', compact('data'));
     }
 
-    public function salesInvoiceData(Request $request)
-    {
-
-        $salesInvoice = SalesInvoice::findOrFail($request['sales_invoice_id']);
-
-        $view = view('admin.sales_invoice_return.sales_invoice', compact('salesInvoice'))->render();
-
-        return response()->json(['view' => $view], 200);
-    }
-
     public function store(CreateSalesInvoiceReturnRequest $request)
     {
+//        dd($request->all());
 
         if (!auth()->user()->can('create_sales_invoices_return')) {
             return redirect()->back()->with(['authorization' => 'error']);
         }
 
+        if (!$request->has('items')) {
+            return redirect()->back()->with(['message' => __('sorry,  items required'), 'alert-type' => 'error']);
+        }
+
         try {
+
             DB::beginTransaction();
 
-            $branch_id = $request['branch_id'];
+            $data = $request->all();
 
-            if (!authIsSuperAdmin()) {
-                $branch_id = auth()->user()->branch_id;
+            $invoice_data = $this->salesInvoiceReturn->prepareInvoiceData($data);
+
+            $invoice_data['created_by'] = auth()->id();
+            $invoice_data['branch_id'] = authIsSuperAdmin() ? $request['branch_id'] : auth()->user()->branch_id;
+
+            $salesInvoiceReturn = SalesInvoiceReturn::create($invoice_data);
+
+            $this->salesInvoiceReturn->salesInvoiceReturnTaxes($salesInvoiceReturn, $data);
+
+            foreach ($data['items'] as $item) {
+
+                $item_data = $this->salesInvoiceReturn->calculateItemTotal($item, $data['invoice_type']);
+
+                $item_data['sales_invoice_return_id'] = $salesInvoiceReturn->id;
+
+                $salesInvoiceItemReturn = SalesInvoiceItemReturn::create($item_data);
+
+                if (isset($item['taxes'])) {
+                    $salesInvoiceItemReturn->taxes()->attach($item['taxes']);
+                }
             }
 
-            $sales_invoice = SalesInvoice::findOrFail($request['sales_invoice_id']);
+            if ($salesInvoiceReturn->status == 'finished' && in_array($salesInvoiceReturn->invoice_type, ['direct_invoice', 'direct_sale_quotations'])) {
 
-            $customer_id = null;
+                $acceptQuantityData = $this->handleQuantityServices->acceptQuantity($salesInvoiceReturn->items, 'pull');
 
-            if ($sales_invoice->customer) {
-                $customer_id = $sales_invoice->customer->id;
-            }
+                if (isset($acceptQuantityData['status']) && !$acceptQuantityData['status']) {
 
-            $sales_invoice_return_data = $this->prepareInvoiceData($request, $branch_id);
+                    $message = isset($acceptQuantityData['message']) ? $acceptQuantityData['message'] : __('sorry, please try later');
 
-            $last_invoice = SalesInvoiceReturn::where('branch_id', $branch_id)->latest('created_at')->first();
-
-            $sales_invoice_return_data['invoice_number'] = $last_invoice ? $last_invoice->invoice_number + 1 : 1;
-
-            $sales_invoice_return_data['customer_id'] = $customer_id;
-
-            $sales_invoice_return = SalesInvoiceReturn::create($sales_invoice_return_data);
-
-            $this->handlePointsLog($sales_invoice_return);
-
-            foreach ($request['return_part_ids'] as $index => $part_id) {
-
-                $sales_invoice_item = SalesInvoiceItems::findOrFail($request['sales_invoice_items_id_' . $index]);
-
-                if ($request['return_qty_' . $index] > $sales_invoice_item->sold_qty) {
-
-                    return redirect()->back()->with(['message' => __('words.sorry return quantity is more than sold'),
-                        'alert-type' => 'error']);
+                    return redirect()->back()->with(['message' => $message, 'alert-type' => 'error']);
                 }
-
-                $sales_invoice_return_item = $this->calculateItemTotal($request, $part_id, $index);
-
-                $sales_invoice_return_item['sales_invoice_return_id'] = $sales_invoice_return->id;
-
-                $return_item = SalesInvoiceItemReturn::create($sales_invoice_return_item);
-
-                $purchase_invoice = $sales_invoice_item->purchaseInvoice;
-
-                if (!$purchase_invoice) {
-                    return redirect()->back()->with(['message' => __('words.something-wrong'),
-                        'alert-type' => 'error']);
-                }
-
-                $invoice_item = $purchase_invoice->items()->where('part_id', $part_id)->first();
-
-                $this->affectedPurchaseItem($invoice_item, $request['return_qty_' . $index]);
-
-                $this->affectedPart($part_id, $request['return_qty_' . $index], $request['selling_price_' . $index]);
             }
 
             DB::commit();
         } catch (\Exception $e) {
             DB::rollBack();
-//            dd($e->getMessage());
 
-            return redirect()->back()->with(['message' => __('words.try-again'),
-                'alert-type' => 'error']);
+            dd($e->getMessage());
+
+            return redirect()->back()->with(['message' => __('words.sales-invoice-return-cant-created'), 'alert-type' => 'error']);
         }
 
-        $url = route('admin:expenseReceipts.create', ['sales_invoice_return_id' => $sales_invoice_return->id]);
-
-        try {
-
-            $this->sendNotification('return_sales_invoice', 'customer',
-                [
-                    'sales_invoice_return' => $sales_invoice_return,
-                    'message' => 'New sales invoice return created for you please check'
-                ]);
-
-            if ($sales_invoice_return->customer && $sales_invoice_return->customer->email) {
-
-                $this->sendMail($sales_invoice_return->customer->email, 'sales_invoice_return_status', 'sales_invoice_return_create',
-                    'App\Mail\SalesInvoiceReturn');
-            }
-
-        } catch (\Exception $e) {
-
-            return redirect($url)->with(['message' => __('words.sale-return-invoice-created'), 'alert-type' => 'success']);
-        }
-
-        return redirect($url)->with(['message' => __('words.sale-return-invoice-created'), 'alert-type' => 'success']);
-//        return Response::json($url, 200);
-    }
-
-    public function expensesReceipts(SalesInvoiceReturn $invoice)
-    {
-        return view('admin.sales_invoice_return.expenses_receipts.index', compact('invoice'));
+        return redirect(route('admin:sales.invoices.return.index'))
+            ->with(['message' => __('words.sale-invoice-return-created'), 'alert-type' => 'success']);
     }
 
     public function show(Request $request)
@@ -337,37 +312,32 @@ class SalesInvoiceReturnController extends Controller
 
     public function destroy(SalesInvoiceReturn $invoice)
     {
-
         if (!auth()->user()->can('delete_sales_invoices_return')) {
             return redirect()->back()->with(['authorization' => 'error']);
         }
 
-        $invoice->delete();
-
-        $this->sendNotification('return_sales_invoice', 'customer',
-            [
-                'sales_invoice_return' => $invoice,
-                'message' => 'your sales invoice return deleted, please check'
-            ]);
-
-        if ($invoice->customer && $invoice->customer->email) {
-
-            $this->sendMail($invoice->customer->email, 'sales_invoice_return_status', 'sales_invoice_return_delete',
-                'App\Mail\SalesInvoiceReturn');
+        if ($invoice->status == 'finished') {
+            return redirect()->back()->with(['message' => __('words.sale-return-invoice-cant-deleted'), 'alert-type' => 'error']);
         }
+
+        $invoice->delete();
 
         return redirect()->back()->with(['message' => __('words.sale-return-invoice-deleted'), 'alert-type' => 'success']);
     }
 
     public function deleteSelected(Request $request)
     {
+
         if (!auth()->user()->can('delete_sales_invoices_return')) {
             return redirect()->back()->with(['authorization' => 'error']);
         }
 
         try {
+
             if (isset($request->ids) && is_array($request->ids)) {
-                SalesInvoiceReturn::whereIn('id', array_unique($request->ids))->delete();
+
+                SalesInvoiceReturn::whereIn('id', array_unique($request->ids))->where('status','!=', 'finished')->delete();
+
                 return redirect()->back()
                     ->with([
                         'message' => __('words.selected-row-deleted'),
@@ -440,17 +410,18 @@ class SalesInvoiceReturnController extends Controller
         }
     }
 
-    public function selectSalesInvoiceOrReturnedReceipt(Request $request) {
+    public function selectSalesInvoiceOrReturnedReceipt(Request $request)
+    {
 
         $rules = [
             'type' => 'required|string|in:normal,direct_invoice,direct_sale_quotations,from_sale_quotations,from_sale_supply_order',
             'item_id' => 'required|integer|min:0'
         ];
 
-        if ( in_array($request['type'], ['direct_invoice', 'direct_sale_quotations'])) {
+        if (in_array($request['type'], ['direct_invoice', 'direct_sale_quotations'])) {
             $rules['item_id'] = 'required|integer|exists:sales_invoices,id';
 
-        }else {
+        } else {
             $rules['item_id'] = 'required|integer|exists:returned_sale_receipts,id';
         }
 
@@ -465,10 +436,10 @@ class SalesInvoiceReturnController extends Controller
 
             if (in_array($request['type'], ['direct_invoice', 'direct_sale_quotations'])) {
 
-                $returnedItem =  SalesInvoice::with('items')->where('id', $request['item_id'])
+                $returnedItem = SalesInvoice::with('items')->where('id', $request['item_id'])
                     ->where('status', 'finished')
                     ->first();
-            }else {
+            } else {
 
                 $returnedItem = ReturnedSaleReceipt::find($request['item_id']);
             }
@@ -483,7 +454,7 @@ class SalesInvoiceReturnController extends Controller
 
             $view = view('admin.sales_invoice_return.returned_items', compact('returnedItem'))->render();
 
-            return response()->json(['view' => $view, 'index'=> $index], 200);
+            return response()->json(['view' => $view, 'index' => $index], 200);
 
         } catch (\Exception $e) {
 
@@ -537,4 +508,75 @@ class SalesInvoiceReturnController extends Controller
         return redirect()->back()->with(['message' => __('purchase.return.terms.successfully'), 'alert-type' => 'success']);
     }
 
+    /**
+     * @param Builder $items
+     * @return mixed
+     */
+    private function dataTableColumns(Builder $items)
+    {
+        $viewPath = 'admin.sales_invoice_return.datatables.options';
+
+        return DataTables::of($items)->addIndexColumn()
+
+            ->addColumn('date', function ($item) use ($viewPath) {
+                $withDate = true;
+                return view($viewPath, compact('item', 'withDate'))->render();
+            })
+            ->addColumn('branch_id', function ($item) use ($viewPath) {
+                $withBranch = true;
+                return view($viewPath, compact('item', 'withBranch'))->render();
+            })
+            ->addColumn('number', function ($item) {
+                return $item->number;
+            })
+            ->addColumn('type', function ($item) {
+                return __($item->type);
+            })
+            ->addColumn('clientable_type', function ($item) use ($viewPath) {
+                return class_basename($item->clientable_type);
+            })
+            ->addColumn('clientable_id', function ($item) use ($viewPath) {
+                return $item->clientable ? $item->clientable->name : '---';
+            })
+            ->addColumn('total', function ($item) use ($viewPath) {
+                return $item->total;
+            })
+            ->addColumn('status', function ($item) use ($viewPath) {
+                $withStatus = true;
+                return view($viewPath, compact('item', 'withStatus'))->render();
+            })
+            ->addColumn('created_at', function ($item) {
+                return $item->created_at->format('y-m-d h:i:s A');
+            })
+            ->addColumn('updated_at', function ($item) {
+                return $item->updated_at->format('y-m-d h:i:s A');
+            })
+            ->addColumn('action', function ($item) use ($viewPath) {
+                $withActions = true;
+                return view($viewPath, compact('item', 'withActions'))->render();
+
+            })->addColumn('options', function ($item) use ($viewPath) {
+                $withOptions = true;
+                return view($viewPath, compact('item', 'withOptions'))->render();
+            })->rawColumns(['action'])->rawColumns(['actions'])->escapeColumns([])->make(true);
+    }
+
+    private function filter(Request $request, Builder $data): Builder
+    {
+        return $data->where(function ($query) use ($request) {
+
+            if ($request->filled('branch_id')) {
+                $query->where('branch_id', $request->branch_id);
+            }
+
+            if ($request->filled('number')) {
+                $query->where('number', $request->number);
+            }
+
+            if ($request->filled('type')) {
+                $query->where('type', $request->type);
+            }
+
+        });
+    }
 }
